add AvatarUpload component and integrate avatar upload functionality in AvatarForm and AccountPage
